Metadata before compiling and its handling
This adds default values to default metadata and the option to set it
before compiling the protocol (i.e. while planning to TO).
Additionally, the (dynamic) metadata fields are handled correctly in the
ProtocolForm and metadata is shown while still planning.
Additionally, protocol.start_time is used instead of
protocoltype.usual_time.

/close #47
/close #65
/close #106