Add default protocol recurrence
If a meeting has a regular interval (e.g. weekly or biweekly), setting
the protocoltype recurrence interval simplifies generating regular
protocols.
If recurrence is set and no [sitzung;…]-tag is given in the protocol, a
new protocol will be created on date+recurrence days on protocol upload,
provided this date is still in the future.